requirements.txt
numpy
flask
airsim
SpeechRecognition
pocketsphinx
transformers
torch
pydub
opencv-python-headless
msgpack-rpc-python
accelerate
bitsandbytes
pillow
requests
torchvision